feat: #254 답변 좋아요/댓글 개수 조회 방식 변경

+++ src/main/java/com/server/capple/domain/answerComment/dto/AnswerCommentResponse.java
@@ -27,7 +27,7 @@ public class AnswerCommentResponse {
         private Long answerCommentId;
         private Long writerId;
         private String content;
-        private Long heartCount;
+        private Integer heartCount;
         private LocalDateTime createdAt;
 
     }
